+++ l2rpn_baselines/LeapNetEncoded/__init__.py
@@ -1,0 +1,11 @@
+__all__ = [
+    "LeapNetEncoded",
+    "evaluate",
+    "train",
+    "LeapNetEncoded_NN"
+]
+
+from l2rpn_baselines.LeapNetEncoded.LeapNetEncoded import LeapNetEncoded
+from l2rpn_baselines.LeapNetEncoded.evaluate import evaluate
+from l2rpn_baselines.LeapNetEncoded.train import train
+from l2rpn_baselines.LeapNetEncoded.LeapNetEncoded_NN import LeapNetEncoded_NN