+++ CHANGELOG.rst
@@ -4,6 +4,18 @@ Change Log
 --------
 - stack multiple states in `utils/DeepQAgent`
 
+[0.5.0] - 2020-08-??
+--------------------
+- [FIXED] the counting of the action types frequency in tensorboard (for some baselines)
+- [FIXED] a broken Replay buffer `utils.ReplayBuffer` (used in some baselines)
+- [FIXED] a bug in using multiple environments for some baselines
+- [FIXED] wrong q value update for some baselines
+- [IMPROVED] descriptions and computation of the tensorboard information (for some baselines)
+- [IMPROVED] performance optimization for training and usage of some baselines
+- [ADDED] better serializing as json of the `utils.NNParam` class
+- [ADDED] the LeapNetEncoded baselines that uses a leap neural network (leap net) to create an
+  embedding of the state of the powergrid.
+
 [0.4.4] - 2020-07-07
 --------------------
 - [FIXED] now the baselines can fully support the grid2op MultiMix environment.